feat: Users 엔티티 수정 및 자동매매 구현

+++ src/main/java/com/traders/traders/module/users/domain/Users.java
@@ -43,7 +43,7 @@ public class Users extends AuditTime implements UserDetails {
 	private int leverage;
 
 	@Column(nullable = false)
-	private long quantity;
+	private double quantity;
 
 	@Column
 	private String binanceApiKey;
