+++ src/main/java/com/traders/traders/module/users/domain/repository/dao/AutoTradingSubscriberDao.java
@@ -7,12 +7,12 @@ import lombok.Getter;
 @Getter
 public class AutoTradingSubscriberDao {
 	private final int leverage;
-	private final long quantity;
+	private final double quantity;
 	private final String binanceApiKey;
 	private final String binanceSecretKey;
 
 	@QueryProjection
-	public AutoTradingSubscriberDao(int leverage, long quantity, String binanceApiKey, String binanceSecretKey) {
+	public AutoTradingSubscriberDao(int leverage, double quantity, String binanceApiKey, String binanceSecretKey) {
 		this.leverage = leverage;
 		this.quantity = quantity;
 		this.binanceApiKey = binanceApiKey;
